+++ apps/mull-ui/src/app/components/expandable-text/expandable-text.scss
