+++ apps/mull-ui/src/app/components/index.ts
@@ -1,5 +1,7 @@
 export * from './event-page-header/event-page-header';
 export * from './event-page-info/event-page-info';
+export * from './expandable-text/expandable-text';
 export * from './header/header';
 export * from './mull-back-button/mull-back-button';
+export * from './mull-button/mull-button';
 export * from './navigation-bar/navigation-bar';
