+++ libs/ui-lib/src/index.ts
@@ -1,5 +1,3 @@
-export * from './components/expandable-text/expandable-text';
-export * from './components/mull-button/mull-button';
 export * from './lib/custom-time-picker/custom-time-picker';
 export * from './lib/custom-text-input/custom-text-input';
 export * from './lib/pill-options/pill-options';
